add registerDevice endpoint to Eliza
